feat: allow `corgyparser` functions to specify nargs
BREAKING CHANGE: By default, arguments with a custom parser will now
always receive a single string argument, regardless of the argument
type. This behavior can be controlled with the newly added `nargs`
argument for `@corgyparser`. Setting this argument (which is `None` by
default) will cause the set value to be passed as the `nargs` argument
for `ArgumentParser.add_argument`, and the correspondingly, the custom
parser will be called with all arguments in a list (unless `nargs` is
`None`).