feat: allow `Corgy` attributes with default values to be unset
BREAKING CHANGE: `Corgy` now uses defaults only during `__init__`.
This means after init, attributes with default values are
undistinguishable from the rest. The key effect of this change is that
an attribute with a default value can be unset using `del`. Previously,
such attributes would "fallback" to the default value.