feat: make parsing of `Optional` attributes consistent with other types
BREAKING CHANGE: Attributes marked with `Optional` are no longer added
by `Corgy.add_args_to_parser` with `optional=True`, unless they have a
default value. This behavior is consistent with the treatment of other
types--`Optional` only indicates that the attribute can be `None`. In
line with this change, `Optional` attributes are added with a custom
action with allows arguments to accept zero arguments (which will be
treated as `None`).